refactor: Отделены утилитарные функции офферов от функций точек
