+++ src/view/utils/observer.js
@@ -12,7 +12,6 @@ export default class Observer {
   }
 
   _notify(event, payload) {
-    debugger
     this._observers.forEach((observer) => observer(event, payload));
   }
 }
